move entry_fetcher defines in header
